fix: hydrate chaindata & miniMetadata before subscribing to balances
