+++ apps/extension/src/core/domains/tokens/handler.ts
@@ -3,7 +3,7 @@ import { CustomErc20Token, CustomErc20TokenCreate } from "@core/domains/tokens/t
 import { talismanAnalytics } from "@core/libs/Analytics"
 import { ExtensionHandler } from "@core/libs/Handler"
 import { chaindataProvider } from "@core/rpcs/chaindata"
-import { miniMetadataUpdater } from "@core/rpcs/mini-metadata-updater"
+import { updateAndWaitForUpdatedChaindata } from "@core/rpcs/mini-metadata-updater"
 import { Port, RequestIdOnly } from "@core/types/base"
 import { assert } from "@polkadot/util"
 import { githubUnknownTokenLogoUrl } from "@talismn/chaindata-provider"
@@ -25,25 +25,8 @@ export default class TokensHandler extends ExtensionHandler {
       // token handlers -----------------------------------------------------
       // --------------------------------------------------------------------
       case "pri(tokens.subscribe)": {
-        await miniMetadataUpdater.hydrateFromChaindata()
-
-        await chaindataProvider.hydrateChains()
-        await chaindataProvider.hydrateEvmNetworks()
-
-        const chains = await chaindataProvider.chainsArray()
-        const { statusesByChain } = await miniMetadataUpdater.statuses(chains)
-        const goodChains = [...statusesByChain.entries()].flatMap(([chainId, status]) =>
-          status === "good" ? chainId : []
-        )
-        await chaindataProvider.hydrateTokens(goodChains)
-
-        const [chainIds, evmNetworkIds] = await Promise.all([
-          chaindataProvider.chainIds(),
-          chaindataProvider.evmNetworkIds(),
-        ])
-
         // TODO: Run this on a timer or something instead of when subscribing to tokens
-        await miniMetadataUpdater.update(chainIds, evmNetworkIds)
+        await updateAndWaitForUpdatedChaindata()
 
         // triggers a pending scan if any
         // doing this here as this is the only place where we hydrate tokens from github
